+++ src/event_handler.rs
@@ -5,10 +5,10 @@ use crate::config::keymap::{build_override_table, OverrideEntry};
 use crate::config::keymap_action::KeymapAction;
 use crate::config::modmap_operator::{Interruptable, Keys, ModmapOperator, MultiPurposeKey, PressReleaseKey};
 use crate::config::nested_remap::Remap;
+use crate::config::Config;
 use crate::device::InputDeviceInfo;
 use crate::event::{Event, KeyEvent, RelativeEvent};
 use crate::operator_handler::OperatorHandler;
-use crate::Config;
 use evdev::KeyCode as Key;
 use lazy_static::lazy_static;
 use log::{debug, warn};
